requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.1.0
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
requests-mock>=1.11.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Connector dependencies
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.97.0
msal>=1.24.0
psycopg2-binary>=2.9.7

# Core dependencies
airweave-sdk>=0.3.0,<0.4.0

# Data processing
dagster>=0.26.18
dagster-webserver>=0.26.18
dagster-postgres>=0.26.18
pyarrow>=14.0.1
numpy<2.0.0
watchdog>=3.0.0

# AI/LLM
openai>=1.3.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
tqdm>=4.65.0

# Monitoring
prometheus_client>=0.16.0

# Vector store dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
numpy<2.0.0

# Text processing
spacy>=3.7.2
python-magic>=0.4.27
pypdf2>=3.0.0

# Database
redis>=5.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
